# Multiple optimizations (speed and memory usage improvement)