--- a/README.md
+++ b/README.md
@@ -18,12 +18,15 @@ operations in __O(1)__. The number of memory allocations is reduced to the very 
 It requires __at least__ 8 bytes for each cell in the grid, in other words,
 more than 8 Mb of memory for a 1000x1000 gridmap.
 
-# How fast is it?
+## How fast is it?
 
 Even if a fair amount of work was spent (mostly for fun) to optimize this software, it is still an A*
 algorithm, that will be outperformed by other, more advanced, algorithms.
 
 If you are looking for state-of-the-art path finders, you might be interested to www.movingai.com
+
+This particularly implementation is efficient but also easy to read and understand.
+
 
 ## Usage