15% speed improvement when using the default heuristics.