@facontidavide 2019-10-07
It is actually faster now

Old benchmark
```
BM_AStar_Big          129443995 ns    129443311 ns           32
BM_AStar_Smooth_1000   91320707 ns     91319632 ns           46
BM_AStar_Small          5938943 ns      5938910 ns          706
```

New benchmark
```
BM_AStar_Big           94683926 ns     94681340 ns           73
BM_AStar_Smooth_1000   65355006 ns     65354110 ns          107
BM_AStar_Small          4174486 ns      4174392 ns         1676
```