@chataign 2019-10-12
I didn't check all the changes in detail but it looks good and is faster. I merged a change from master to only clear the cells that were visited, I found that it makes a big difference also. I created tag v1 with the previous version before the merge. Thanks for all the work!!